backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.31
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.2
tqdm==4.66.1
apscheduler==3.10.4
python-dotenv==1.0.0
# Auth deps
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0
pydantic[email]==2.5.0
email-validator==2.1.0
